+++ src/ts/langs/de.ts
@@ -1,0 +1,20 @@
+/*!
+ |  German Translation
+ |  @author     SamBrishes <sam@pytes.net> (https://www.pytes.net)
+ |  @source     None
+ */
+import RatSelect from "rat.select";
+
+RatSelect.Strings["de"] = {
+    buttonAll: "Alle",
+    buttonNone: "Keine",
+    disabled: "Dieses Feld ist deaktiviert",
+    empty: "Keine Optionen verfügbar",
+    multiple: "Wähle eine oder mehrere Optionen...",
+    multipleCount: (count) => {
+        return `[0] ${count === 1? "Option": "Optionen"} ausgewählt...`
+    },
+    multipleLimit: "Keine weiteren Optionen wählbar",
+    search: "Tippen um zu suchen...",
+    single: "Wähle eine Option..."
+};