+++ src/ts/langs/de_AT.ts
@@ -1,6 +1,0 @@
-
-import RatSelect from "rat.select";
-
-RatSelect.Strings["de_AT"] = {
-    "key": "value"
-};