+++ src/ts/langs/de_DE.ts
@@ -1,6 +1,0 @@
-
-import RatSelect from "rat.select";
-
-RatSelect.Strings["de_DE"] = {
-    "key": "value"
-};