+++ src/ts/langs/el.ts
@@ -1,0 +1,18 @@
+/*!
+ |  Greek Translation
+ |  @author     AGTGreg (https://github.com/AGTGreg)
+ |  @source     https://github.com/pytesNET/tail.select/pull/151
+ */
+import RatSelect from "rat.select";
+
+RatSelect.Strings["el"] = {
+    buttonAll: "Όλα",
+    buttonNone: "Κανένα",
+    disabled: "Απενεργοποιημένο",
+    empty: "Δεν υπάρχουν διαθέσιμες επιλογές",
+    multiple: "Διαλέξτε εώς :limit Επιλογές...",
+    multipleCount: "Διαλέξτε εώς :limit Επιλογές...",
+    multipleLimit: "Δεν υπάρχει δυνατότητα άλλης επιλογής",
+    search: "Αναζήτηση...",
+    single: "Επιλογή",
+};