+++ src/ts/langs/fa_IR.ts
@@ -1,0 +1,18 @@
+/*!
+ |  Farsi Translation
+ |  @author     abdulhalim (https://github.com/abdulhalim)
+ |  @source     https://github.com/pytesNET/rat.select/issues/129
+ */
+import RatSelect from "rat.select";
+
+RatSelect.Strings["fa_IR"] = {
+    buttonAll: "همه",
+    buttonNone: "هیچکدام",
+    disabled: "کادر غیرفعال است",
+    empty: "انتخابی موجود نیست",
+    multiple: "یک گزینه ای را انتخاب کنید",
+    multipleCount: "حداکثر انتخاب تا:[0] را محدود کنید ...",
+    multipleLimit: "نمیتوانید گزینه‌های بیشتری داشته باشید",
+    search: "برای جستجو تایپ کنید...",
+    single: "یک گزینه ای را انتخاب کنید"
+};