+++ src/ts/langs/fi.ts
@@ -1,0 +1,18 @@
+/*!
+ |  Finnish Translation
+ |  @author     Noxludio (https://github.com/noxludio)
+ |  @source     https://github.com/pytesNET/rat.select/pull/35
+ */
+import RatSelect from "rat.select";
+
+RatSelect.Strings["fi"] = {
+    buttonAll: "Kaikki",
+    buttonNone: "Ei mitään",
+    disabled: "Kenttä on poissa käytöstä",
+    empty: "Ei vaihtoehtoja",
+    multiple: "Valitse...",
+    multipleCount: "Valitse maksimissaan [0]...",
+    multipleLimit: "Muita vaihtoehtoja ei voi valita",
+    search: "Hae tästä...",
+    single: "Valitse..."
+};