+++ src/ts/langs/it.ts
@@ -1,0 +1,18 @@
+/*!
+ |  Italian Translation
+ |  @author     Alberto Vincenzi (https://github.com/albertovincenzi)
+ |  @source     https://github.com/pytesNET/rat.select/issues/43
+ */
+import RatSelect from "rat.select";
+
+RatSelect.Strings["it"] = {
+    buttonAll: "Tutti",
+    buttonNone: "Nessuno",
+    disabled: "Questo Campo è disabilitato",
+    empty: "Nessuna voce disponibile",
+    multiple: "Seleziona una Voce...",
+    multipleCount: "Selezione limitata a [0] Voci...",
+    multipleLimit: "Non puoi selezionare più Voci",
+    search: "Digita per cercare...",
+    single: "Seleziona una Voce..."
+};