+++ src/ts/langs/pl.ts
@@ -1,0 +1,18 @@
+/*!
+ |  Polish Translation
+ |  @author     funbu (https://github.com/funnbu)
+ |  @source     https://github.com/pytesNET/tail.select/issues/162
+ */
+import RatSelect from "rat.select";
+
+RatSelect.Strings["pl"] = {
+    buttonAll: "Wszystkie",
+    buttonNone: "Żadne",
+    disabled: "To pole jest zablokowane",
+    empty: "Brak dostępnych opcji",
+    multiple: "Zaznacz opcję",
+    multipleCount: "Zaznacz maksymalnie [0] opcji",
+    multipleLimit: "Nie możesz zaznaczyć więcej opcji",
+    search: "Wyszukaj...",
+    single: "Zaznacz opcję"
+};